SCAL-319970 Separate per-user token/org state into its own Durable Object
The keep-warm token store + active-org state were bolted onto
ConversationStorageServerSQLite, conflating two unrelated concerns: ephemeral
per-conversation message streaming (30-min TTL) and durable per-user token/org
state (11h keep-warm + 14-day idle TTL). The shared alarm() even had to branch
on which kind of instance it was.

Split the token/org concern into a new Durable Object, UserTokenStoreSQLite:
- New DO class + USER_TOKEN_OBJECT binding (migration v8 in wrangler.jsonc).
- Owns active-org, active-org-token, token-store, and /touch, plus the
  keep-warm refresh alarm and 14-day idle abandonment. Its alarm() is now
  single-purpose (no branching).
- ConversationStorageServerSQLite goes back to pure conversation streaming; its
  alarm() is solely the 30-min cleanup.
- StorageServiceClient now takes a second namespace and routes token/org calls to
  USER_TOKEN_OBJECT while conversation calls stay on CONVERSATION_STORAGE_OBJECT.
- getStorageService passes both bindings.

No data migration needed: this is pre-prod, and the token store self-seeds from
props on next connect while the active org defaults to the user's current org —
so a fresh (empty) UserTokenStore instance simply re-populates on reconnect.

Tests: token/org tests moved to a dedicated user-token-store-server.spec.ts
(targeting the new DO), with added active-org routing cases. The conversation
spec is now token-free. Full suite green (718 passing).
