+++ worker-configuration.d.ts
@@ -4,12 +4,13 @@
 declare namespace Cloudflare {
 	interface GlobalProps {
 		mainModule: typeof import("./src/index");
-		durableNamespaces: "ThoughtSpotMCP";
+		durableNamespaces: "ThoughtSpotMCP" | "UserTokenStoreSQLite";
 	}
 	interface Env {
 		OAUTH_KV: KVNamespace;
 		MCP_OBJECT: DurableObjectNamespace<import("./src/index").ThoughtSpotMCP>;
 		CONVERSATION_STORAGE_OBJECT: DurableObjectNamespace<import("./src/index").ConversationStorageServerSQLite>;
+		USER_TOKEN_OBJECT: DurableObjectNamespace<import("./src/index").UserTokenStoreSQLite>;
 		ANALYTICS: AnalyticsEngineDataset;
 		ASSETS: Fetcher;
 	}
